requirements.txt
streamlit
openai
python-dotenv
requests
beautifulsoup4
pandas
